Split seasonal_rendering.rs into module directory (1532 -> ~400 lines each)
Break the monolithic seasonal_rendering.rs (1532 lines) into a directory
module with focused sub-files:

- constants.rs (70 lines): tuning constants for all effects
- types.rs (236 lines): SeasonalEffectsConfig, SeasonalRenderingState, Saveable impls
- compute.rs (207 lines): pure helper functions for intensity calculations
- system.rs (147 lines): ECS system and plugin
- tests_compute.rs (383 lines): tests for seasonal compute + cell counting
- tests_weather.rs (263 lines): tests for rain/snow/storm/lightning
- tests_types.rs (272 lines): tests for types, saveable, constants, combos
- mod.rs (33 lines): thin facade re-exporting all public items

All files are under 500 lines. External callers are unaffected since
mod.rs re-exports everything at the same paths.

Part of #1389
